statestorage: Remove go-sqlite3 dependency
statestorage does not build because the package
"github.com/mattn/go-sqlite3" is not found with the "_" in front
of it.
However, if package is installed first, go says it is imported and not
used by the file statestorage_mgr.go
It seems like it can be removed since database/sql is used instead?