statestorage: Refactor and fix InitDB for new DB case
When running the command with 3 parameters to create a new database, and
the file does NOT exist previously, then the code for InitDb still
incorrectly went into the final else clause to print:

"DB /data/statestorage.db must exist, or else the statestorage
manager must be started also with a filename to a JSON pathlist."

It's easy to get the logic wrong because of the complex call interface
to InitDb that uses a boolean parameter (often not a good idea).  This
proposal refactors InitDB into two separate functions, since the two
code paths perform almost completely different tasks, and I believe it
fixes the logical error.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>